+++ Changelogs/Changelog.md
@@ -1,6 +1,7 @@
 ### Added
 - See the characters with the most progress for a given achievement in its tooltip
 - Additional options and data management
+- Option to ignore filters in the Watch List
 
 ### Changed
 - Custom objectives are now handled in a different way to allow more flexibility in the future